[Bronze III] Title: Tax, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/14182. Tax/Tax.cc
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
 	int money;
-	long long result;
+	int result;
 	while (true)
 	{
 		cin >> money;
